Calculate active fire properties based on NASA FIRMS polygons

This function allows to efficiently extract the properties of fire events occurred in the region of interest from the NASA FIRMS active fire polygon datasets. For each polygon, the fire events properties like fire pixel brightness temperature, and fire radiative power (frp) along with fire hotspots for the desired year is returned. The required resources for this indicator are:

nasa_firms

Usage

calc_active_fire_properties()

Value

A function that returns a tibble with a column for the 15 different fire events variables including lon/lat coordinates.

Examples

# \dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
# }
# \dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) %>%
  read_sf() %>%
  get_resources(get_nasa_firms(years = 2021, instrument = "VIIRS")) %>%
  calc_indicators(calc_active_fire_properties()) %>%
  tidyr::unnest(active_fire_properties)
#> Resource 'nasa_firms' is already available.

aoi
#> Simple feature collection with 21 features and 20 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -71.80933 ymin: 18.57668 xmax: -71.33201 ymax: 18.69931
#> Geodetic CRS:  WGS 84
#> # A tibble: 21 × 21
#>    WDPAID NAME  DESIG_ENG ISO3  assetid bright_ti4  scan track acq_date acq_time
#>     <dbl> <chr> <chr>     <chr>   <int>      <dbl> <dbl> <dbl> <chr>       <int>
#>  1 478140 Sier… National… DOM         1       338.  0.49  0.41 2021-01…     1812
#>  2 478140 Sier… National… DOM         1       341.  0.39  0.36 2021-01…     1754
#>  3 478140 Sier… National… DOM         1       298.  0.67  0.73 2021-01…      524
#>  4 478140 Sier… National… DOM         1       299.  0.67  0.73 2021-01…      524
#>  5 478140 Sier… National… DOM         1       333.  0.33  0.55 2021-02…     1718
#>  6 478140 Sier… National… DOM         1       345.  0.68  0.74 2021-02…     1700
#>  7 478140 Sier… National… DOM         1       346.  0.42  0.45 2021-03…     1736
#>  8 478140 Sier… National… DOM         1       346.  0.41  0.61 2021-03…     1712
#>  9 478140 Sier… National… DOM         1       336.  0.39  0.36 2021-03…     1754
#> 10 478140 Sier… National… DOM         1       331.  0.39  0.36 2021-03…     1754
#> # ℹ 11 more rows
#> # ℹ 11 more variables: satellite <chr>, instrument <chr>, confidence <chr>,
#> #   version <int>, bright_ti5 <dbl>, frp <dbl>, daynight <chr>, type <int>,
#> #   longitude <dbl>, latitude <dbl>, geom <POLYGON [°]>
# }