Portfolio methods for mapme.biodiversity

mapme_options() sets default options for mapme.biodiversity to control the behavior of downstream functions. Mainly, the output path as well as the temporal directory for intermediate files can be set. Additionally, the verbosity can be set. The testing options should not be set by users, as it controls the behavior of the package during automated test pipelines. Might be extended by other options in the future.

get_resources() data sets required for the calculation of indicators can be made available. The function supports the specification of several resource functions. To determine the output path, temporary directory and verbosity, the output of mapme_options() is used.

calc_indicators() calculates specific biodiversity indicators. A requirement is that the resources that are mandatory inputs for the requested indicators are available locally. Multiple indicators and their respective additional arguments can be supplied.

This function reads and crops available resources to the extent of a single asset. Specific resources can be queried. If not supplied (the default), all available resources will be prepared.

Usage

mapme_options(..., outdir, chunk_size, verbose, aria_bin, testing, log_dir)

get_resources(x, ...)

calc_indicators(x, ...)

prep_resources(x, avail_resources = NULL, resources = NULL)

Arguments

...

One or more functions for resources/indicators

outdir

A length one character indicating the output path.

chunk_size

A numeric of length one giving the maximum chunk area in ha. Defaults to 100,000 ha.

verbose

A logical, indicating if informative messages should be printed.

aria_bin

A character vector to an aria2c executable for parallel downloads.

testing

A logical. Not to be set by users. Controls the behavior during automated test pipelines.

log_dir

A character path pointing toward a GDAL-writable destination used to log erroneous assets. Defaults to NULL, meaning that erroneous assets will not be serialized to disk. If specified, a GPKG named file.path(log_dir, paste0(Sys.Date(), "_mapme-error-assets.gpkg")) will be created and appended to in case of erroneous assets.

x

An sf object with features of type "POLYGON"

avail_resources

A list object of available resources. If NULL (the default), the available resources will automatically be determined.

resources

A character vector with the resources to be prepared. If it it is NULL (the default) all available resources will be prepared.

Value

mapme_options() returns a list of options if no arguments are specified. Otherwise sets matching arguments to new values in the package's internal environment.

get_resources() is called for its side effect of making resources available in the package environment. Returns x, invisibly.

calc_indicators() returns x, invisibly, with an additional nested list column per requested indicator.

prep_resources() returns a list with prepared vector and raster resources as sf and SpatRaster-objects.

Examples

library(mapme.biodiversity)
mapme_options()
#> $outdir
#> [1] "/tmp/Rtmp6gbTM8/mapme-data"
#> 
#> $chunk_size
#> [1] 1e+05
#> 
#> $verbose
#> [1] FALSE
#> 
#> $aria_bin
#> NULL
#> 
#> $testing
#> [1] FALSE
#> 
#> $log_dir
#> NULL
#>