NASADEM HGT v001

This resource is processed by the Land Processes Distributed Active Archive Center (LP DAAC) and made available at the Microsoft Planetery Computer. NASADEM are distributed in 1 degree latitude by 1 degree longitude tiles and consist of all land between 60° N and 56° S latitude. This accounts for about 80% of Earth’s total landmass.

Usage

get_nasa_srtm()

Source

https://planetarycomputer.microsoft.com/dataset/nasadem

Value

A function that returns an sf footprint object.

References

NASA JPL (2020). NASADEM Merged DEM Global 1 arc second V001. NASA EOSDIS Land Processes DAAC. Accessed 2023-07-01 from doi:10.5067/MEaSUREs/NASADEM/NASADEM_HGT.001