Calculate maximum temperature statistics

This function allows to efficiently calculate maximum temperature statistics from Worldclim for polygons. For each polygon, the desired statistic/s (min, max, sum, mean, median, sd or var) is/are returned.

Usage

calc_temperature_max_wc(engine = "extract", stats = "mean")

Arguments

engine

The preferred processing functions from either one of "zonal", "extract" or "exactextract" as character.

stats

Function to be applied to compute statistics for polygons either single or multiple inputs as character. Supported statistics are: "mean", "median", "sd", "min", "max", "sum" "var".

Value

A tibble with a column for maximum temperature statistics (in °C)

Details

The required resources for this indicator are:

maximum temperature layer from worldclim_max_temperature

Examples

# \dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
# }
# \dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) %>%
  read_sf() %>%
  get_resources(get_worldclim_max_temperature(years = 2018)) %>%
  calc_indicators(
    calc_temperature_max_wc(
      engine = "extract",
      stats = c("mean", "median")
    )
  ) %>%
  tidyr::unnest(temperature_max_wc)

aoi
#> Simple feature collection with 12 features and 8 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -71.80933 ymin: 18.57668 xmax: -71.33201 ymax: 18.69931
#> Geodetic CRS:  WGS 84
#> # A tibble: 12 × 9
#>    WDPAID NAME          DESIG_ENG ISO3  assetid tmax_mean tmax_median date      
#>     <dbl> <chr>         <chr>     <chr>   <int>     <dbl>       <dbl> <date>    
#>  1 478140 Sierra de Ne… National… DOM         1      20.8        20.6 2018-01-01
#>  2 478140 Sierra de Ne… National… DOM         1      20.5        20.3 2018-02-01
#>  3 478140 Sierra de Ne… National… DOM         1      22.1        22.0 2018-03-01
#>  4 478140 Sierra de Ne… National… DOM         1      22.5        22.4 2018-04-01
#>  5 478140 Sierra de Ne… National… DOM         1      21.4        21.3 2018-05-01
#>  6 478140 Sierra de Ne… National… DOM         1      22.7        22.7 2018-06-01
#>  7 478140 Sierra de Ne… National… DOM         1      23.1        23.1 2018-07-01
#>  8 478140 Sierra de Ne… National… DOM         1      23.8        23.7 2018-08-01
#>  9 478140 Sierra de Ne… National… DOM         1      23.1        23.0 2018-09-01
#> 10 478140 Sierra de Ne… National… DOM         1      22.4        22.3 2018-10-01
#> 11 478140 Sierra de Ne… National… DOM         1      22.1        21.9 2018-11-01
#> 12 478140 Sierra de Ne… National… DOM         1      21.3        21.2 2018-12-01
#> # ℹ 1 more variable: geom <POLYGON [°]>
# }