Calculate emission statistics

This functions allows to efficiently calculate emission statistics for areas of interest. For each year in the analysis timeframe, the forest losses from Hansen et al. (2013) are overlayed with the respective emission layer from Harris et al. (2021) and area-wise emission statistics are calculated for each year.

Usage

calc_treecoverloss_emissions(years = 2000:2023, min_size = 10, min_cover = 35)

Arguments

years

A numeric vector with the years for which to calculate emissions caused by treecover loss.

min_size

The minimum size of a forest patch in ha.

min_cover

The minimum threshold of stand density for a pixel to be considered forest in the year 2000.

Value

A function that returns a tibble with a column for years and emissions (in Mg).

Details

The required resources for this indicator are:

gfw_treecover

gfw_lossyear

gfw_emissions

Examples

# \dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
# }
# \dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) %>%
  read_sf() %>%
  get_resources(
    get_gfw_treecover(version = "GFC-2023-v1.11"),
    get_gfw_lossyear(version = "GFC-2023-v1.11"),
    get_gfw_emissions()
  ) %>%
  calc_indicators(
    calc_treecoverloss_emissions(years = 2016:2017, min_size = 1, min_cover = 30)
  ) %>%
  portfolio_long()
#> Resource 'gfw_treecover' is already available.
#> Resource 'gfw_lossyear' is already available.
#> Resource 'gfw_emissions' is already available.

aoi
#> Simple feature collection with 2 features and 10 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -71.80933 ymin: 18.57668 xmax: -71.33201 ymax: 18.69931
#> Geodetic CRS:  WGS 84
#> # A tibble: 2 × 11
#>   WDPAID NAME  DESIG_ENG ISO3  assetid indicator datetime   variable unit  value
#>    <dbl> <chr> <chr>     <chr>   <int> <chr>     <date>     <chr>    <chr> <dbl>
#> 1 478140 Sier… National… DOM         1 treecove… 2016-01-01 emissio… Mg    4296.
#> 2 478140 Sier… National… DOM         1 treecove… 2017-01-01 emissio… Mg    4970.
#> # ℹ 1 more variable: geom <POLYGON [°]>
# }