Calculate accessibility statistics

Accessibility is the ease with which larger cities can be reached from a certain location. This function allows to efficiently calculate accessibility statistics (i.e. travel time to nearby major cities) for polygons. For each polygon, the desired statistic/s (mean, median or sd) is/are returned. The required resources for this indicator are:

nelson_et_al

Format

A tibble with a column for accessibility statistics (in minutes)

Details

The following arguments can be set:

stats_accessibility

Function to be applied to compute statistics for polygons either one or multiple inputs as character. Supported statistics are: "mean", "median", "sd", "min", "max", "sum" "var".

engine

The preferred processing functions from either one of "zonal", "extract" or "exactextract" as character.

Examples

# \dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
# }
# \dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) %>%
  read_sf() %>%
  init_portfolio(
    years = 2022,
    outdir = outdir,
    tmpdir = tempdir(),
    add_resources = FALSE,
    verbose = FALSE
  ) %>%
  get_resources("nelson_et_al",
    range_traveltime = c("5k_10k", "100k_200k", "500k_1mio", "1mio_5mio")
  ) %>%
  calc_indicators("traveltime", stats_accessibility = c("min", "max"), engine = "extract") %>%
  tidyr::unnest(traveltime)

aoi
#> Simple feature collection with 4 features and 8 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -71.80933 ymin: 18.57668 xmax: -71.33201 ymax: 18.69931
#> Geodetic CRS:  WGS 84
#> # A tibble: 4 × 9
#>   WDPAID NAME           DESIG_ENG ISO3  assetid minutes_min minutes_max distance
#>    <dbl> <chr>          <chr>     <chr>   <int>       <dbl>       <dbl> <chr>   
#> 1 478140 Sierra de Nei… National… DOM         1          35         348 5k_10k  
#> 2 478140 Sierra de Nei… National… DOM         1         162         528 100k_20…
#> 3 478140 Sierra de Nei… National… DOM         1         239         553 500k_1m…
#> 4 478140 Sierra de Nei… National… DOM         1         106         464 1mio_5m…
#> # ℹ 1 more variable: geom <POLYGON [°]>
# }